Fix a simple docker utility
Fixes "/root/entrypoint.sh: line 7: ./kdoki/kokorodoki: No such file or directory" and "Error: Error querying device -1." After fixing those, the docker build was hanging on daemon mode, which is not really desired to simply download models. Also, it is not desirable to git clone in the dockerfile when you are using the container to test and debug.